fix build and behavior
true-calc failed to build after previous commit because of
`sprintf` inclusion (it didn't fit); moreover, there were
some missing _fall-through comments_.

Note: added a flag to disable what's called _strict aliasing_. Not
really sure what it is, but it produces a lot of errors and seems to
be some optimization feature.

Also fix constness of button titles and a few typos.